% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_codonSplitCpp}
\alias{cpp_codonSplitCpp}
\title{Split a string into codons}
\usage{
cpp_codonSplitCpp(cDNA)
}
\arguments{
\item{cDNA}{(char) A single string}
}
\value{
(char) a vector of triplets
}
\description{
Split a string into codons
}
\examples{
# Split a string into codons
cpp_codonSplitCpp("ATGAAATGA")
}
\author{
Boris Steipe, \email{boris.steipe@utoronto.ca}
ORCID: 0000-0002-1134-6758
License: (c) Author (2018) + MIT
Date: 2018-12-31
}
