% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptPlus.R
\docType{package}
\name{rptPlus}
\alias{rptPlus}
\alias{rptPlus-package}
\title{rptPlus: R Package Template - enhanced}
\description{
The "rptPlus.R" file is a dummy function that contains
roxygen2 tags in the header which add the required directives for
calling compiled code to the NAMESPACE file.
}
