% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRptApp.R
\name{runRptApp}
\alias{runRptApp}
\title{\code{runRptApp} run the shiny app \code{./inst/shiny-scripts/rptApp/}.}
\usage{
runRptApp()
}
\value{
NULL (invisible) Invoked for its side-effect of launching a shiny app.
}
\description{
\code{runRptApp} run the shiny app \code{./inst/shiny-scripts/rptApp/}.
}
\examples{
# No runnable example applies - but BiocCheck() requires one. So ...
NULL
\dontrun{
runRptApp()
}
}
