% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptPlus.R
\docType{package}
\name{rptPlus}
\alias{rptPlus}
\alias{rptPlus-package}
\title{rptPlus: R Package Template with extars.}
\description{
The rptPlus template package includes support for shiny and Rcpp functions.
The "rptPlus.R" file is a dummy function whose purpose it is to contain
a header that makes roxygen2 add the required directives that
call compiled code to the NAMESPACE file.
}
